pyproject.toml
[tool.poetry]
name = "pupalink_bot"
version = "0.1.0"
description = "Pupalink-Bot which uses the pupalink library, for educational purposes"
authors = ["billaids <jimmy.nelle@hsw-stud.de>"]
license = "GPL-3.0 License"

[tool.poetry.dependencies]
python = "^3.9"
pupalink = "^0.1.0-alpha.4"
Pyrogram = "^1.2.9"
motor = "^2.4.0"
loguru = "^0.5.3"
PyYAML = "^5.4.1"
TgCrypto = "^1.2.2"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
